requirements.txt
discord.py
requests
jsbeautifier
js2py
python-dotenv